Codefactor fixes
